requirements.txt
praw
python-dotenv
openai>=1.0.0